+++ src/survey/survey.resolver.ts
@@ -15,21 +15,39 @@ export class SurveyResolver{
     
     @Query((returns => Survey), {name: 'surveyById'})
     getSurveyById(@Args({name:'surveyId', type:()=>Int}) id:number){
+        if(!id){
+            throw new Error("설문지 id가 입력되지 않았습니다.");
+        }
         return this.surveyService.findSurveyById(id);
     }
 
     @Mutation((returns => String), {name: 'deleteSurvey'})
     deleteSurvey(@Args({name:'surveyId', type:()=>Int}) id:number){
+        if(!id){
+            throw new Error("설문지 id가 입력되지 않았습니다.");
+        }
         return this.surveyService.deleteSurvey(id);
     }
 
     @Mutation((returns => String), {name: 'addSurvey'})
     addSurvey(@Args("addSurveyArgs") survey: AddSurveyArgs){
+        if(survey.title.length === 0){
+            throw new Error("설문지 title이 입력되지 않았습니다.");
+        }
+        if(survey.description.length === 0){
+            throw new Error("설문지 description이 입력되지 않았습니다.");
+        }
         return this.surveyService.addSurvey(survey);
     }
 
     @Mutation((returns => String), {name: 'updateSurvey'})
     upadteSurvey(@Args("updateSurveyArgs") survey: UpdateSurveyArgs){
+        if(survey.title.length === 0){
+            throw new Error("설문지 title이 입력되지 않았습니다.");
+        }
+        if(survey.description.length === 0){
+            throw new Error("설문지 description이 입력되지 않았습니다.");
+        }
         return this.surveyService.upadteSurvey(survey);
     }
 }
